add Dao find order asc and desc + test

+++ src/main/java/tp_group1/spring_boot_pokemon/model/Attack.java
@@ -1,25 +1,22 @@
 package tp_group1.spring_boot_pokemon.model;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 
 @Entity
 public class Attack {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String attackName;
     private String type;
     private int damage;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,4 +43,5 @@ public class Attack {
     public void setDamage(int damage) {
         this.damage = damage;
     }
+
 }
