+++ src/main/java/tp_group1/spring_boot_pokemon/model/Species.java
@@ -1,28 +1,23 @@
 package tp_group1.spring_boot_pokemon.model;
 
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 
 @Entity
 public class Species {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private int id;
+    private Long id;
 
     private String name;
     private String type;
     private int initialHealthPoints;
 
-
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
